megabase: Redesign science blueprint
The new design uses internal mirroring to ensure proper lane balance,
whereas the previous design did not. The new design is also additionally
optimized to yield an additional 40 SPM, which is then doubled due to
the mirroring to get a total of 880 SPM. One additional benefit of the
mirrored design is that some belt inputs can be shared between the
mirrored blocks, which makes routing the inputs easier.