megabase: Add new integrated blueprint for circuits
The blueprint produces roughly twice the output of the “Modules &
Circuits” blueprint, and its output is tailored to the ratios needed to
do non-military infinite research. The design is built around consuming
720 copper plates per second, which is exactly 16 blue belts.

The most complicated part of the design is the belt logistics for copper
plates. Depending on which outputs are currently demanded, it is
necessary for the allocation of the copper to change in response in
order to achieve maximum throughput. The current design tries to strike
a balance between space usage, material cost, and graceful handling of
arbitrary demand patterns; the precise set of solutions was arrived at
by a great deal of empirical trial and error.